authorization implemented

+++ src/services/notes/controller/notes-controller.js
@@ -3,36 +3,42 @@ import response from "../../../utils/response.js";
 import InvariantError from "../../../exceptions/invariant-error.js";
 import NotFoundError from "../../../exceptions/not-found-error.js";
 import NoteRepositories from "../repositories/note-repositories.js";
+import AuthorizationError from '../../../exceptions/authorization-error.js';
 
-export const createNote = async(req, res, next) => {
+
+export const createNote = async (req, res, next) => {
     const { title, body, tags } = req.validated;
+    const { id: owner } = req.user;
     const note = await NoteRepositories.createNote({
         title,
         body,
         tags,
+        owner
     });
-    
     if (!note) {
-        return next(new InvariantError('catatan gagal ditamhkan'));
+        return next(new InvariantError('Catatan gagal ditambahkan'));
     }
-
     return response(res, 201, 'Catatan berhasil ditambahkan', note);
 };
 
 export const getNotes = async (req, res) => {
-  const notes = await NoteRepositories.getNotes();
+  const { id: owner } = req.user;
+  const notes = await NoteRepositories.getNotes(owner);
   return response(res, 200, 'Catatan sukses ditampilkan', notes);
 };
 
-export const getNotesById = async (req, res, next) => {
-    const { id } = req.params;
-    const note = await NoteRepositories.getNoteByid(id);
-
-    if (!note) {
-        return next(new NotFoundError('Catatan tidak ditemukan'));
-    }
-
-    return response (res, 200, 'Catatan sukses ditampilkan', note);
+export const getNoteById = async (req, res, next) => {
+  const { id } = req.params;
+  const { id: owner } = req.user;
+  const isOwner = await NoteRepositories.verifyNoteOwner(id, owner);
+  if (!isOwner) {
+    return next(new AuthorizationError('Anda tidak berhak mengakses resource ini'));
+  }
+  const note = await NoteRepositories.getNoteById(id);
+  if (!note) {
+    return next(new NotFoundError('Catatan tidak ditemukan'));
+  }
+  return response(res, 200, 'Catatan sukses ditampilkan', note);
 };
 
 export const editNote = async (req, res, next) => {
@@ -58,13 +64,39 @@ export const editNote = async (req, res, next) => {
 };
 
 export const deleteNoteById = async (req, res, next) => {
-        const { id } = req.params;
-        const deleteNote = await NoteRepositories.deleteNote(id);
+  const { id } = req.params;
+  const { id: owner } = req.user;
+  const isOwner = await NoteRepositories.verifyNoteOwner(id, owner);
+  if (!isOwner) {
+    return next(new AuthorizationError('Anda tidak berhak mengakses resource ini'));
+  }
+  const deletedNote = await NoteRepositories.deleteNote(id);
+  if (!deletedNote) {
+    return next(new NotFoundError('Catatan tidak ditemukan'));
+  }
+  return response(res, 200, 'Catatan berhasil dihapus', deletedNote);
+};
 
-
-        if (!deleteNote) {
-            return next(new NotFoundError('Catatan tidak ditemukan'));
-        }
-
-        return response(res, 200, 'Catatan berhasil dihapus', deleteNote);
+export const editNoteById = async (req, res, next) => {
+  const { id } = req.params;
+  const {
+    title,
+    body,
+    tags
+  } = req.validated;
+  const { id: owner } = req.user;
+  const isOwner = await NoteRepositories.verifyNoteOwner(id, owner);
+  if (!isOwner) {
+    return next(new AuthorizationError('Anda tidak berhak mengakses resource ini'));
+  }
+  const note = await NoteRepositories.editNote({
+    id,
+    title,
+    body,
+    tags
+  });
+  if (!note) {
+    return next(new NotFoundError('Catatan tidak ditemukan'));
+  }
+  return response(res, 200, 'Catatan berhasil diperbarui', note);
 };
